Cleanup and added code todos (#242)
* added todos, removed unused code

* renames

* move code from h to cpp for qslim

* moved code for remeshing from h to cpp

* moved code from h to cpp

* moved rational.h to utils,
naming convention fixes in tetwild
