+++ app/tetwild/TriangleInsertion.cpp
@@ -1,10 +1,10 @@
-#include "Rational.hpp"
+#include <wmtk/utils/Rational.hpp>
 #include "TetWild.h"
 
 #include <wmtk/utils/Delaunay.hpp>
 #include "common.h"
 #include "wmtk/TetMesh.h"
-#include "wmtk/auto_table.hpp"
+#include "wmtk/TetMeshCutTable.hpp"
 #include "wmtk/utils/GeoUtils.h"
 #include "wmtk/utils/InsertTriangleUtils.hpp"
 #include "wmtk/utils/Logger.hpp"
@@ -145,7 +145,7 @@ auto internal_insert_single_triangle(
     };
 
     const auto& [flag, intersected_tets, intersected_edges, intersected_pos] =
-        wmtk::triangle_insert_prepare_info<apps::Rational>(
+        wmtk::triangle_insert_prepare_info<wmtk::Rational>(
             m,
             face,
             marked_tet_faces, // output
